add runtime check
