+++ packages/cli/src/controller/generate-controller.spec.ts
@@ -6,7 +6,7 @@ import os from 'os';
 import path from 'path';
 import {EventFragment, FunctionFragment} from '@ethersproject/abi/src.ts/fragments';
 import {DEFAULT_TS_MANIFEST, NETWORK_FAMILY} from '@subql/common';
-import type {
+import {
   SubqlRuntimeDatasource as EthereumDs,
   EthereumLogFilter,
   EthereumDatasourceKind,
@@ -23,7 +23,6 @@ import {
   replaceArrayValueInTsManifest,
   resolveToAbsolutePath,
   splitArrayString,
-  tsStringify,
 } from '../utils';
 import {
   constructDatasourcesTs,
@@ -38,6 +37,7 @@ import {
   removeKeyword,
   tsExtractor,
   yamlExtractor,
+  tsStringify,
 } from './generate-controller';
 
 const mockConstructedFunctions: SelectedMethod[] = [
@@ -56,7 +56,7 @@ const mockConstructedEvents: SelectedMethod[] = [
 
 const mockDsFn = (): EthereumDs[] => [
   {
-    kind: 'ethereum/Runtime' as EthereumDatasourceKind.Runtime,
+    kind: EthereumDatasourceKind.Runtime,
     startBlock: 1,
     options: {
       abi: 'Erc721',
@@ -70,14 +70,14 @@ const mockDsFn = (): EthereumDs[] => [
       handlers: [
         {
           handler: 'handleTransaction',
-          kind: 'ethereum/TransactionHandler' as EthereumHandlerKind.Call,
+          kind: EthereumHandlerKind.Call,
           filter: {
             function: 'approve(address,uint256)',
           },
         },
         {
           handler: 'handleLog',
-          kind: 'ethereum/LogHandler' as EthereumHandlerKind.Event,
+          kind: EthereumHandlerKind.Event,
           filter: {
             topics: ['Transfer(address,address,uint256)'],
           },
@@ -143,7 +143,7 @@ describe('CLI codegen:generate', () => {
     const expectedAsset = new Map();
     expectedAsset.set('Erc721', {file: './abis/erc721.json'});
     expect(constructedDs).toStrictEqual({
-      kind: 'ethereum/Runtime' as EthereumDatasourceKind.Runtime,
+      kind: EthereumDatasourceKind.Runtime,
       startBlock: 1,
       options: {
         abi: 'Erc721',
@@ -155,14 +155,14 @@ describe('CLI codegen:generate', () => {
         handlers: [
           {
             handler: 'handleTransferFromErc721Tx',
-            kind: 'ethereum/TransactionHandler' as EthereumHandlerKind.Call,
+            kind: EthereumHandlerKind.Call,
             filter: {
               function: 'transferFrom(address,address,uint256)',
             },
           },
           {
             handler: 'handleApprovalErc721Log',
-            kind: 'ethereum/LogHandler' as EthereumHandlerKind.Event,
+            kind: EthereumHandlerKind.Event,
             filter: {
               topics: ['Approval(address,address,uint256)'],
             },
@@ -557,7 +557,7 @@ describe('CLI codegen:generate', () => {
       address: 'aaa',
     };
     expect(constructDatasourcesYaml(mockUserInput)).toStrictEqual({
-      kind: 'ethereum/Runtime' as EthereumDatasourceKind.Runtime,
+      kind: EthereumDatasourceKind.Runtime,
       startBlock: 1,
       options: {
         abi: 'Erc721',
@@ -569,14 +569,14 @@ describe('CLI codegen:generate', () => {
         handlers: [
           {
             handler: 'handleTransferFromErc721Tx',
-            kind: 'ethereum/TransactionHandler' as EthereumHandlerKind.Call,
+            kind: EthereumHandlerKind.Call,
             filter: {
               function: 'transferFrom(address,address,uint256)',
             },
           },
           {
             handler: 'handleApprovalErc721Log',
-            kind: 'ethereum/LogHandler' as EthereumHandlerKind.Event,
+            kind: EthereumHandlerKind.Event,
             filter: {
               topics: ['Approval(address,address,uint256)'],
             },
