+++ packages/cli/src/utils/utils.ts
@@ -14,7 +14,6 @@ import {
   DEFAULT_MANIFEST,
   DEFAULT_TS_MANIFEST,
 } from '@subql/common';
-import type {SubqlRuntimeHandler as EthSubqlRuntimeHandler} from '@subql/types-ethereum';
 import axios from 'axios';
 import cli, {ux} from 'cli-ux';
 import ejs from 'ejs';
@@ -173,36 +172,6 @@ export function replaceArrayValueInTsManifest(content: string, key: string, newV
 export function extractArrayValueFromTsManifest(content: string, key: string): string | null {
   const [startIndex, endIndex] = findArrayIndicesTsManifest(content, key);
   return content.slice(startIndex, endIndex + 1);
-}
-// @ts-ignore
-// TODO
-
-export function tsStringify(
-  obj: EthSubqlRuntimeHandler | EthSubqlRuntimeHandler[] | string,
-  indent = 2,
-  currentIndent = 0
-): string {
-  if (typeof obj !== 'object' || obj === null) {
-    if (typeof obj === 'string' && obj.includes('EthereumHandlerKind')) {
-      return obj; // Return the string as-is without quotes
-    }
-    return JSON.stringify(obj);
-  }
-
-  if (Array.isArray(obj)) {
-    const items = obj.map((item) => tsStringify(item, indent, currentIndent + indent));
-    return `[\n${items.map((item) => ' '.repeat(currentIndent + indent) + item).join(',\n')}\n${' '.repeat(
-      currentIndent
-    )}]`;
-  }
-
-  const entries = Object.entries(obj);
-  const result = entries.map(([key, value]) => {
-    const valueStr = tsStringify(value, indent, currentIndent + indent);
-    return `${' '.repeat(currentIndent + indent)}${key}: ${valueStr}`;
-  });
-
-  return `{\n${result.join(',\n')}\n${' '.repeat(currentIndent)}}`;
 }
 
 export function extractFromTs(
